// src/common/enums/createPromo.enum.ts
export enum DiscountMetric {
  PERCENTAGE = 'percentage',
  AMOUNT = 'amount',
}

export enum DiscountCut {
  UNMUDL = 'unmudl',
  BOTH = 'both',
  GIFT = 'gift',
}

export enum ApplyTo {
  ALL = 'all',
  SELECTED = 'selected',
}

export enum PromoDuration {
  FOREVER = 'forever',
  ONCE = 'once',
}

export enum Status {
  ACTIVE = 'active',
  SUSPENDED = 'suspended',
}